pyproject.toml
[tool.poetry]
name = "gas-estimation-example"
version = "0.1.0"
description = ""
authors = ["Peteris Erins <peteris@auditless.com>"]

[tool.poetry.dependencies]
python = "3.9.13"
click = "^8.0.3"
python-dotenv = "^0.19.2"
"starknet.py" = "^0.2.4-alpha.0"
cairo-nile = "^0.6.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
